Feature #8

+++ features/weatherInfo/appServices/src/main/java/com/diskin/alon/pagoda/weatherinfo/appservices/model/HourForecastDto.kt
@@ -1,5 +1,5 @@
 package com.diskin.alon.pagoda.weatherinfo.appservices.model
 
-data class HourForecastDto(val hour: Int,
+data class HourForecastDto(val hour: Long,
                            val condition: WeatherConditionDto,
                            val temp: Double)
